// src/app/modules/league-of-legends/modules/models/lol-game-detais.model.ts
// Generated by https://quicktype.io

export class GameDetails {
  esportsGameId?: string;
  esportsMatchId?: string;
  gameMetadata?: GameMetadata;
  frames: Frame[] = [];
}

export class Frame {
  rfc460Timestamp: string = '';
  gameState: GameState = GameState.InGame;
  blueTeam: Team = new Team();
  redTeam: Team = new Team();
}

export class Team {
  totalGold: number = 0;
  inhibitors: number = 0;
  towers: number = 0;
  barons: number = 0;
  totalKills: number = 0;
  dragons: Dragon[] = [];
  participants: Participant[] = [];
}

export enum Dragon {
  Elder = 'elder',
  Infernal = 'infernal',
  Mountain = 'mountain',
  Ocean = 'ocean',
}

export interface Participant {
  participantId: number;
  totalGold: number;
  level: number;
  kills: number;
  deaths: number;
  assists: number;
  creepScore: number;
  currentHealth: number;
  maxHealth: number;
}

export enum GameState {
  Finished = 'finished',
  InGame = 'in_game',
}

export interface GameMetadata {
  patchVersion: string;
  blueTeamMetadata: TeamMetadata;
  redTeamMetadata: TeamMetadata;
}

export interface TeamMetadata {
  esportsTeamId: string;
  participantMetadata: ParticipantMetadatum[];
}

export interface ParticipantMetadatum {
  participantId: number;
  esportsPlayerId: string;
  summonerName: string;
  championId: string;
  role: string;
}